✨ Fold registration into the login page
Entering an email on /login now either logs in or signs up: credential
accounts get the optional password field as before; everything else
receives a sign-in OTP that signs into an existing account or creates
one on verification. Known and unknown emails are indistinguishable, so
the combined flow leaks nothing.

- /register (and its verify page, form, and action) collapse into the
  login equivalents; the route remains as a redirect for old links
- the OTP send happens in a login server action that enforces Turnstile;
  the widget is revealed only when an OTP send is needed, so password
  logins never see it
- when registration is disabled the action falls back to
  email-verification OTPs, which better-auth silently skips for unknown
  emails — no accounts can be created and the verify copy keeps its
  "if an account exists" hedge
