+++ apps/web/public/locales/en/app.json
@@ -40,7 +40,6 @@
   "allDay": "All day",
   "allMembers": "All",
   "allParticipants": "All Participants",
-  "alreadyHaveAccount": "Already have an account? <a>Log in</a>",
   "alreadyLoggedInContinue": "Continue",
   "alreadyLoggedInDescription": "Continue to your dashboard or sign out to use a different account.",
   "alreadyLoggedInTitle": "You're already signed in",
@@ -248,7 +247,6 @@
   "emailChangeVerifyError": "Verification failed. Please try again.",
   "emailDoesNotMatch": "The email address does not match.",
   "emailMismatch": "Email does not match the account email",
-  "emailNotAllowed": "This email is not allowed.",
   "emailPlaceholder": "jessie.smith@example.com",
   "enabled": "Enabled",
   "endOfList": "You've reached the end of the list",
@@ -380,8 +378,9 @@
   "locationPlaceholder": "Joe's Coffee Shop",
   "login": "Login",
   "loginDescription": "Login to your account to continue",
-  "loginFooter": "Don't have an account? <a>Sign up</a>",
   "loginNotConfigured": "Login is currently not configured.",
+  "loginOrSignUpDescription": "Log in to your account or create a new one",
+  "loginOtpSendError": "Something went wrong. Please try again.",
   "loginTitle": "Welcome",
   "loginVerifyCheckEmail": "If an account exists with this email, you will receive a verification code at <b>{email}</b>",
   "loginVerifyTitle": "Finish Logging In",
@@ -542,11 +541,6 @@
   "quickCreateWhyCreateAnAccount": "Why create an account?",
   "refreshLicense": "Refresh License",
   "register": "Register",
-  "registerDescription": "Streamline your scheduling process and save time",
-  "registerOtpSendError": "Something went wrong. Please try again.",
-  "registerTitle": "Create Your Account",
-  "registerVerifyDescription": "Enter the 6-digit code we sent to <b>{email}</b>",
-  "registerVerifyTitle": "Finish Signing Up",
   "remove": "Remove",
   "removeImage": "Remove",
   "removeLicense": "Remove License",
@@ -727,6 +721,8 @@
   "verificationCodePlaceholder": "Enter your 6-digit code",
   "verify": "Verify",
   "verifyEmail": "Verify your email",
+  "verifyEmailDescription": "Enter the 6-digit code we sent to <b>{email}</b>",
+  "verifyEmailTitle": "Verify Your Email",
   "viewPoll": "View poll",
   "viewResults": "View results",
   "weekView": "Week view",
