+++ apps/web/src/app/[locale]/(auth)/login/page.tsx
@@ -14,11 +14,9 @@ import {
   AuthPageContainer,
   AuthPageContent,
   AuthPageDescription,
-  AuthPageExternal,
   AuthPageHeader,
   AuthPageTitle,
 } from "../components/auth-page";
-import { LinkWithRedirectTo } from "../components/link-with-redirect-to";
 import { AuthErrors } from "./components/auth-errors";
 import { LoginWithEmailForm } from "./components/login-email-form";
 
@@ -92,6 +90,13 @@ export default async function LoginPage(props: {
               i18nKey="loginNotConfigured"
               defaults="Login is currently not configured."
             />
+          ) : isRegistrationEnabled ? (
+            <Trans
+              t={t}
+              ns="app"
+              i18nKey="loginOrSignUpDescription"
+              defaults="Log in to your account or create a new one"
+            />
           ) : (
             <Trans
               t={t}
@@ -129,18 +134,6 @@ export default async function LoginPage(props: {
         </div>
       </AuthPageContent>
       <AuthErrors />
-      {isRegistrationEnabled ? (
-        <AuthPageExternal>
-          <Trans
-            t={t}
-            i18nKey="loginFooter"
-            defaults="Don't have an account? <a>Sign up</a>"
-            components={{
-              a: <LinkWithRedirectTo className="text-link" href="/register" />,
-            }}
-          />
-        </AuthPageExternal>
-      ) : null}
     </AuthPageContainer>
   );
 }
