--- apps/web/src/app/[locale]/(auth)/register/verify/components/register-otp-form.tsx
@@ -1,145 +1,0 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { posthog } from "@rallly/posthog/client";
-import { Button } from "@rallly/ui/button";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormMessage,
-} from "@rallly/ui/form";
-import { useSearchParams } from "next/navigation";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-
-import { InputOTP } from "@/components/input-otp";
-import { Trans, useTranslation } from "@/i18n/client";
-import { authClient } from "@/lib/auth-client";
-import { validateRedirectUrl } from "@/lib/utils/redirect";
-
-const otpFormSchema = z.object({
-  otp: z.string().length(6),
-});
-
-export function RegisterOTPForm({ email }: { email: string }) {
-  const { t } = useTranslation();
-  const form = useForm({
-    defaultValues: {
-      otp: "",
-    },
-    resolver: zodResolver(otpFormSchema),
-  });
-
-  const searchParams = useSearchParams();
-  const handleSubmit = form.handleSubmit(async (data) => {
-    // Creates an account if the email is new, otherwise signs into the
-    // existing one. A guest session is linked to the account either way.
-    const res = await authClient.signIn.emailOtp({
-      email,
-      otp: data.otp,
-    });
-
-    if (res.error) {
-      posthog?.capture("register:otp_verify_error", {
-        error_code: res.error.code ?? "UNKNOWN",
-        status: res.error.status,
-      });
-
-      switch (res.error.code) {
-        case "INVALID_OTP":
-          form.setError("otp", {
-            message: t("wrongVerificationCode", {
-              defaultValue: "Your verification code is incorrect",
-            }),
-          });
-          return;
-        case "OTP_EXPIRED":
-          form.setError("otp", {
-            message: t("expiredVerificationCode", {
-              defaultValue:
-                "This code has expired. Request a new one to continue.",
-            }),
-          });
-          return;
-        case "TOO_MANY_ATTEMPTS":
-          form.setError("otp", {
-            message: t("tooManyVerificationAttempts", {
-              defaultValue:
-                "Too many incorrect attempts. Request a new code to continue.",
-            }),
-          });
-          return;
-        default:
-          form.setError("otp", {
-            message: res.error.message,
-          });
-          return;
-      }
-    }
-
-    const redirectTo = validateRedirectUrl(searchParams?.get("redirectTo"));
-
-    // A newly created account has no name yet; send it through onboarding
-    // to pick one up before continuing to its destination.
-    if (!res.data?.user?.name) {
-      const destination = redirectTo
-        ? `/setup?redirectTo=${encodeURIComponent(redirectTo)}`
-        : "/setup";
-      window.location.href = destination;
-      return;
-    }
-
-    window.location.href = redirectTo ?? "/";
-  });
-
-  return (
-    <Form {...form}>
-      <form
-        className="flex flex-col items-center space-y-6 text-center"
-        onSubmit={handleSubmit}
-      >
-        <FormField
-          control={form.control}
-          name="otp"
-          render={({ field }) => {
-            return (
-              <FormItem>
-                <FormControl>
-                  <InputOTP
-                    disabled={
-                      form.formState.isSubmitting ||
-                      form.formState.isSubmitSuccessful
-                    }
-                    autoFocus={true}
-                    onValidCode={() => {
-                      handleSubmit();
-                    }}
-                    {...field}
-                  />
-                </FormControl>
-                <FormDescription className="mt-4">
-                  <Trans i18nKey="verificationCodeHelp" />
-                </FormDescription>
-                <FormMessage />
-              </FormItem>
-            );
-          }}
-        />
-        <Button
-          className="w-full"
-          variant="primary"
-          size="xl"
-          type="submit"
-          loading={
-            form.formState.isSubmitting || form.formState.isSubmitSuccessful
-          }
-        >
-          <Trans i18nKey="continue" defaults="Continue" />
-        </Button>
-      </form>
-    </Form>
-  );
-}
