+++ packages/test-helpers/src/login.ts
@@ -18,7 +18,7 @@ export async function loginWithEmail(
     await page.getByPlaceholder("jessie.smith@example.com").fill(email);
     await page.getByRole("button", { name: "Continue with email" }).click();
     await page
-      .getByRole("heading", { name: "Finish Logging In" })
+      .getByRole("heading", { name: /Verify Your Email|Finish Logging In/ })
       .or(page.getByLabel("Password"))
       .first()
       .waitFor({ timeout: 5000 });
